Preferences and UI fixes.

+++ data/src/main/java/com/example/data/repo/datastore/RemotePlacesDataStore.kt
@@ -28,7 +28,7 @@ class RemotePlacesDataStore @Inject constructor(
     override fun findNearbyPOIs(
         latLng: LatLng
     ): Single<Result<List<SimplePlace>, DataStoreError>> = overpassAPIClient.getPlaces(
-        query = latLng.toOverpassPOIsQueryWithRadius(preferences.placesSearchRadius)
+        query = latLng.toOverpassPOIsQueryWithRadius(preferences.placesSearchRadius.toInt())
     ).result
 
     override fun findNearbyPlacesOfType(
@@ -37,7 +37,7 @@ class RemotePlacesDataStore @Inject constructor(
     ): Single<Result<List<SimplePlace>, DataStoreError>> = overpassAPIClient.getPlaces(
         query = latLng.toOverpassQueryWithRadius(
             query = placeTypeQuery,
-            radius = preferences.placesSearchRadius
+            radius = preferences.placesSearchRadius.toInt()
         )
     ).result
